Moves comment components into seperate folder
